Re-thought and rewrote the the DBIC proxy of the attributes.
Now it applies roles to the attribute metaclass instance in the new()
wrapper. Much cleaner and faster and also works with Moose 1.19.

Unfortunately this does not work with older Moose (a quick glance at the
Moose Changelog did not reveal what is the smallest Moose version that
still works, but anyways, using the latest Moose is not a bad idea).